% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctF.R
\name{correctF}
\alias{correctF}
\title{Temperature dependency of fluorescence offset and levels corrections}
\usage{
correctF(PAM.data,Foffset.a,Foffset.b,Flevel.a, Flevel.b, Fmoffset.a, Fmoffset.b, Fmlevel.a, Fmlevel.b)
}
\arguments{
\item{PAM.data}{preprocessed MONI-PAM data}

\item{Foffset.a}{F' offset correction parameter a}

\item{Foffset.b}{F' offset correction parameter b}

\item{Flevel.a}{F' signal correction parameter a}

\item{Flevel.b}{F' signal correction parameter b}

\item{Fmoffset.a}{Fm' offset correction parameter a}

\item{Fmoffset.b}{Fm' offset correction parameter b}

\item{Fmlevel.a}{Fm' signal correction parameter a}

\item{Fmlevel.b}{Fm' signal correction parameter b}
}
\value{
This function will data table
}
\description{
This is an optional function. This function will correct Fm' and F' value including offset and magnitude, due to the effect of temperature on the measuring light. The corresponding Y(II) and ETR were recalculated and all the F', Fm' or Y(II) were removed when they are less than 0. The correction is based on the equations of fluorescence offset correction - Foffset.a X temp_PAM+Foffset.b and fluorescence signal level correction - Flevel.a X temp_PAM+Flevel.b.
}
