% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diurnalParams.R
\name{diurnalParams}
\alias{diurnalParams}
\title{Estimate diurnal chlorophyll fluorescence parameters}
\usage{
diurnalParams(filtered.data,FmR.data,save.file,save.path)
}
\arguments{
\item{filtered.data}{the final filtered MONI-PAM data retrieved from data filtering step 6  (i.e., flag6.adjacent in Intro_to_LongTermPAM.Rmd)}

\item{filtered.fvfm}{the final FV/FM data retrieved from \link{FindFVFM} function using final filtered data in filter 6 (i.e., flag6.adjacent in Intro_to_LongTermPAM.Rmd). F0 and Fm are used here for ChlF paramters' calculation}

\item{FmR.data}{FmR and F0R data can be retrieved from the \link{ChlFRef} function or defined manually by the user.  This data should be a data frame or data table and include columns 'season','head_tree','head','tree_num','FmR' and 'F0R'}

\item{save.file}{TRUE or FALSE. If TRUE, output data will be saved to local folder via save.path argument}

\item{save.path}{local folder for saving the output data generated from this function}
}
\value{
This function will return a data frame including estimated diurnal parameters and original data
}
\description{
This function estimated diurnal parameters,including quenching parameters - PQ, qLT,NPQT, and NPQr, and yield parameters - Phi_NPQT, Phi_NPQr, and Phi_fD.
}
