% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter5.expand.R
\name{filter.expand.fc}
\alias{filter.expand.fc}
\title{filter logical function for filter5.expand}
\usage{
\method{filter}{expand.fc}(data, f5.fm, f5.fmYII)
}
\description{
filter logical function for filter5.expand
}
