% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.night.R
\name{filter.night.fc}
\alias{filter.night.fc}
\title{filter logic function for filter2.night function}
\usage{
\method{filter}{night.fc}(data, period, f2.Fm, f2.FmYII)
}
\description{
filter logic function for filter2.night function
}
