% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter3.day.R
\name{filter.pm.fc}
\alias{filter.pm.fc}
\title{filter logical function 2 for filter3.day function}
\usage{
\method{filter}{pm.fc}(data, f3.YII, f3.fm)
}
\description{
filter logical function 2 for filter3.day function
}
