% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter4.FVFM.R
\name{filter4.FVFM}
\alias{filter4.FVFM}
\title{Filter MONI-PAM data step 4,remove spurious Fv/Fm (or maximum YII) value for entire observation season}
\usage{
filter4.FVFM(PAM.data,save.path,save.file,f4.Fm=0.15,f4.FmYII=3)
}
\arguments{
\item{PAM.data}{a data.table or data.frame MONI-PAM data generated from \link{filter3.day} function.}

\item{save.path}{local folder for saving your output file}

\item{save.file}{If this argument is set as TRUE, the returned file will be saved to local folder, if FALSE, the file will not be saved into local folder}

\item{f4.Fm}{the threshold of percentage change of Fm between consecutive days. Default value is 0.15, we recommend this argument can be adjusted from 0.05 to 0.3 by an interval of 0.05.}

\item{f4.FmYII}{the threshold of ratio between percentage change in Fm between consecutive days and in Fv/Fm between consecutive days. Default value is 3, we recommend this argument can be adjust between 2 and 5 by an interval of 1.}
}
\value{
\link{filter4.FVFM} will return a data table. Meanwhile, if save.file = TRUE, the output data.table will also be saved into local folder as a 'PAM_Year1_Year2_filter4FvFm.dat' file, where Year1 and Year2 are the minimum and maximum year during this observation season respectively. This output file will contain one new column compared with output file from \link{filter3.day} function named as 'flag4.FvFm'. This column only contain two values: 0 and 1, where 0 means F', Fm' and YII in corresponding row(s) are abnormal data and should be removed from the dataset and 1 means good dataset.
}
\description{
Filter MONI-PAM data step 4,remove spurious Fv/Fm (or maximum YII) value for entire observation season
}
