% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPAMdata.R
\name{formatPAMdata}
\alias{formatPAMdata}
\title{convert MONI-PAM data variables to the correct data format and fill any date or time gaps}
\usage{
formatPAMdata(PAM.data)
}
\arguments{
\item{PAM.data}{The data can be (1) a combined and organized MONI-PAM dataset generated from the \link{readPAM} function or after applying \link{correctF} function, (2) a MONI-PAM dataset after data filtering, generated using filtering functions in this package, such as \link{filter1.NA}, \link{filter2.night} and so on, (3) an FV/FM dataset generated from the \link{FindFVFM} function, (4) a diurnal parameter dataset generated from the \link{diurnalParams} function, or (5) a seasonal parameter dataset generated from the \link{seasonalParams} function.}
}
\value{
\link{formatPAMdata} will generate a MONI-PAM data with all the variables in proper data format for later data filtering, processing and visualization
}
\description{
This function converts MONI-PAM variables to the proper data format and fill any date or time gaps to facilitate data filtering, processing, and visualization.
}
