% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPAM.R
\name{organize.PAMdata}
\alias{organize.PAMdata}
\title{organize data function used only for readPAM function}
\usage{
organize.PAMdata(data, shortnames, i)
}
\description{
organize data function used only for readPAM function
}
