% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCheckFilter.R
\name{plotCheckFilter}
\alias{plotCheckFilter}
\title{MONI-PAM data visualization with data before and after filter}
\usage{
plotCheckFilter(dataAfterFilter, dataBeforeFilter,save.path)
}
\arguments{
\item{dataAfterFilter}{one of MONI-PAM data after data filter (e.g., output file from filters 1 to 6 functions, such as \link{filter3.day} function)}

\item{dataBeforeFilter}{original organized and combined MONI-PAM data which is generated after \link{readPAM} function or other previous filtered data such as output data from \link{filter2.night} function.}

\item{save.path}{local folder for saving the plotted figures generated from this function}
}
\value{
\link{plotCheckFilter} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function visualizes two MONI-PAM datasets - one before filtering (dataBeforeFilter) and one after filter (dataAfterFilter) - including F',FM',Y(II),PAR, and temperature, grouped into 10-day plots for each head. It also displays sunlight times as vertical lines: sunrise and sunsetStart ('indianred'), solarNoon ('goldenrod'), and dawn/dusk ('purple'). In the figure, red, pink, and orange points means filtered or flagged Y(II), Fm', and F' data in data after filter compared to data before filter.
}
