% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiurnQuench.R
\name{plotDiurnQuench}
\alias{plotDiurnQuench}
\title{Diurnal quenching parameters data visualization}
\usage{
plotDiurnQuench(diurnal.para,save.path)
}
\arguments{
\item{diurnal.para}{estimate diurnal parameters generated from \link{diurnalParams} function in this R package}

\item{save.path}{local folder for saving your plotted figures}
}
\value{
\link{plotDiurnQuench} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function visualizes MONI-PAM diurnal quenching parameters (PQ,qLT,NPQT, and NPQr), and PAR and temperature in one figure with 10 days as a group plotting for each head. At the same time, this function will also show  the sunlight time (colorful vertical lines) in the plot.
}
