% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiurnYield.R
\name{plotDiurnYield}
\alias{plotDiurnYield}
\title{Diurnal quantum yield parameters data visualization}
\usage{
plotDiurnYield(diurnal.para,save.path)
}
\arguments{
\item{diurnal.para}{estimated diurnal quantum yield parameters generated from \link{diurnalParams} function in this R package}

\item{save.path}{local folder for saving your plotted figures}
}
\value{
\link{plotDiurnYield} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function will visualize MONI-PAM diurnal quantum yield parameters data (PhiPSII,PhiNPQT,PhifD, and PhiNPQr) and PAR and temperature in one figure with 10 days as a group plotting for each head and the fieldnote will be marked in the figure if the note file is available. At the same time, this function will also show  the sunlight time (colorful vertical lines) in the plot, sunrise and sunsetStart are 'indianred', solarNoon is 'goldenrod', and dawn and dusk are 'purple'
}
