% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFVFM.R
\name{plotFVFM}
\alias{plotFVFM}
\title{MONI-PAM FV/FM data visualization}
\usage{
plotFVFM(PAM.fvfm,plot.title,save.path)
}
\arguments{
\item{PAM.fvfm}{MONI-PAM FV/FM data retrieved from \link{FindFVFM} function in this R package}

\item{plot.title}{any text to describe clearly about your figure content}

\item{save.path}{local folder for saving the plotted figures generated from this function}
}
\value{
This function will not return/show the plotted figures in the end, instead, it will save the plotted figures to your local folder directly.
}
\description{
This function will visualize MONI-PAM FV/FM data including F0,FM,FV/FM,PAR, and temperature in one figure for each head.
}
