% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPAM.R
\name{plotPAM}
\alias{plotPAM}
\title{MONI-PAM data visualization}
\usage{
plotPAM(PAM.data,plot.title,save.path)
}
\arguments{
\item{PAM.data}{a combined and organized MONI-PAM dataset generated from \link{readPAM} function or after applying \link{correctF} function, or a MONI-PAM data after data filtering which is generated using filtering functions in this package, such as \link{filter1.NA}, \link{filter2.night} and so on.}

\item{plot.title}{any text to describe clearly about your figure content}

\item{save.path}{local folder for saving the plotted figures generated from this function}
}
\value{
\link{plotPAM} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function visualizes MONI-PAM data including F',FM',Y(II),PAR, and temperature in a single figure, grouping data into 10-day intervals for each head. It also displays sunlight time using colorful vertical lines: sunrise and sunsetStart in 'indianred', solarNoon in 'goldenrod', and dawn and dusk in 'purple'. These sunlight times will be used in the data filtering process.
}
