% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeasonPara.R
\name{plotSeasonPara}
\alias{plotSeasonPara}
\title{MONI-PAM seasonal parameters data visualization}
\usage{
plotSeasonPara(season.param,save.path)
}
\arguments{
\item{season.param}{seasonal chlorophyll fluorescence parameters retrieved from \link{seasonalParams} function in this R package}

\item{save.path}{local folder for saving the plotted figures generated from this function}
}
\value{
\link{plotSeasonPara} will return/show the plotted seasonal quenching and yield parameters figures in the end, and save the plotted figures to your folder directly.
}
\description{
This function visualize MONI-PAM seasonal quenching (qLs,PQs, and NPQs) and quantum yield (FV/FM, Phi_NPQs, Phi(f,Ds)) parameters for each head.
}
