% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalePAMdata.R
\name{scalePAMdata}
\alias{scalePAMdata}
\title{obtain extreme values of MONI-PAM data}
\usage{
scalePAMdata(PAM.data)
}
\arguments{
\item{PAM.data}{a combined organized MONI-PAM data which is generated from \link{readPAM} function, or a MONI-PAM data after data filtering which is generated from filter function in this package such as \link{filter1.NA}, \link{filter2.night} and so on.}
}
\value{
\link{scalePAMdata} will return a data.table which contains the maximum and minimum values of F',FM',YII, PAR, and temperature values for each plotting groups (every 10 days as a group) and each MONI-head
}
\description{
These extreme values of MONI-PAM data will be used for MONI-PAM data visualization
}
