% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter4.FVFM.R
\name{filter.FvFm.fc41}
\alias{filter.FvFm.fc41}
\title{filter logical function 1 for filter4.FvFm function}
\usage{
\method{filter}{FvFm.fc41}(data, f4.Fm, f4.FmYII)
}
\description{
filter logical function 1 for filter4.FvFm function
}
