% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter3.day.R
\name{filter.o.fc}
\alias{filter.o.fc}
\title{filter logical function 1 for filter3.day function}
\usage{
\method{filter}{o.fc}(data, f3.Fm, f3.FmYII)
}
\description{
filter logical function 1 for filter3.day function
}
