% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.night.R
\name{filter2.night}
\alias{filter2.night}
\title{Filter PAM-PAM data step 2,remove spurious F', Fm' and Y(II) value from the night}
\usage{
filter2.night(PAM.data,f2.Fm=0.03,f2.FmYII=3,save.path,save.file)
}
\arguments{
\item{PAM.data}{a data.table or data.frame PAM-PAM data generated from \link{filter1.lowF} function.}

\item{f2.Fm}{the threshold of percentage change of Fm' between consecutive points. Default value is 0.03, we recommend this argument can be adjusted from 0.01 to 0.1 by an interval of 0.01.}

\item{f2.FmYII}{the threshold of ratio between percentage change in Fm' between consecutive points and in Y(II) between consecutive points. Default value is 3, we recommend this argument can be set as default value or adjusted between 2 and 5 by an interval of 1}

\item{save.path}{local folder for saving your output file}

\item{save.file}{If this argument is set as TRUE, the returned file will be saved to local folder, if FALSE, the file will not be saved into local folder}
}
\value{
\link{filter2.night} will return a data table. Meanwhile, if save.file = TRUE, the output data.table will also be saved into local folder as a 'PAM_Year1_Year2_filter2night.dat' file, where Year1 and Year2 are the minimum and maximum year during this observation season respectively. This output file will contain one new column compared with output file from \link{filter1.lowF} function named as 'flag2.night'. This column only contain two values: 0 and 1, where 0 means F', Fm' and YII in corresponding row(s) are abnormal data and should be removed from the dataset and 1 means good dataset.
}
\description{
Filter PAM-PAM data step 2,remove spurious F', Fm' and Y(II) value from the night
}
