% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPAMdata.R
\name{formatPAMdata}
\alias{formatPAMdata}
\title{convert MONI-PAM data variables to correct data format}
\usage{
formatPAMdata(PAM.data)
}
\arguments{
\item{PAM.data}{a combined organized MONI-PAM data which is generated from \link{readPAM} function, or a MONI-PAM data after data filtering which is generated from filter function in this package such as \link{filter1.lowF}, \link{filter2.night} and so on.}
}
\value{
\link{formatPAMdata} will generate a MONI-PAM data with all the variables in proper data format for later data filtering and processing
}
\description{
This function will convert the MONI-PAM variables to proper data format for easily used for data filtering and processing
}
