% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalParams.R
\name{seasonalParams}
\alias{seasonalParams}
\title{Estimate seasonal chlorophyll fluorescence parameters}
\usage{
seasonalParams(filtered.fvfm,FmR.data,save.file,save.path)
}
\arguments{
\item{fvfm.afterFilter}{the final MONI-PAM fvfm data retrieved from cleaning step 6 (i.e., flag6.fvfm in Intro_to_processMONIPAM.Rmd) by using \link{FindFvFm} function in this R package}

\item{FmR.data}{FmR and F0R data can be retrieved from the \link{ChlFRef} function or defined manually by the user.  This data should be a data frame or data table and include columns 'season','head_tree','head','tree_num','FmR' and 'F0R'}

\item{save.file}{TRUE or FALSE. If TRUE, output data will be saved to local folder via save.path argument}

\item{save.path}{local folder for saving the output data generated from this function}
}
\value{
This function will return a data frame including new estimated seasonal parameters and Fv/Fm data
}
\description{
This function estimates seasonal parameters,including quenching parameters PQs, qLs, and NPQs, and yield parameters Phi_NPQs and Phi_fDs.
}
